📁 여행 코스 생성, 여행지 추천, 지도 페이지 폴더 정리
