+++ src/App.js
@@ -3,21 +3,21 @@ import MainPage from "./pages/MainPage";
 import OptionsPage from "./pages/OptionsPage";
 import SignUpPage from "./pages/SignUpPage";
 import LoginPage from "./pages/LoginPage";
-import DestinationPage from "./pages/DestinationPage";
-import ThemePage from "./pages/ThemePage";
-import PeriodPage from "./pages/PeriodPage";
-import PartyPage from "./pages/PartyPage";
-import ThemePage2 from "./pages/ThemePage2";
-import PeriodPage2 from "./pages/PeriodPage2";
-import ResultPage from "./pages/ResultPage";
-import ChatPage from "./pages/ChatPage";
+import DestinationPage from "./pages/Itinerary/DestinationPage";
+import ThemePage from "./pages/Itinerary/ThemePage";
+import PeriodPage from "./pages/Itinerary/PeriodPage";
+import PartyPage from "./pages/Destination/PartyPage";
+import ThemePage2 from "./pages/Destination/ThemePage2";
+import PeriodPage2 from "./pages/Destination/PeriodPage2";
+import ResultPage from "./pages/Destination/ResultPage";
+import ChatPage from "./pages/Itinerary/ChatPage";
 import ItineraryPage from "./pages/ItineraryPage";
-import MapPage from "./pages/MapPage";
-import MapResultPage from "./pages/MapResultPage";
-import DetailResultPage from "./pages/DetailResultPage";
+import MapPage from "./pages/Map/MapPage";
+import MapResultPage from "./pages/Map/MapResultPage";
+import DetailResultPage from "./pages/Map/DetailResultPage";
 import MyPage from "./pages/MyPage";
 import DetailedDestPage from "./pages/DetailedDestPage";
-import LoadingPage from "./pages/LoadingPage";
+import LoadingPage from "./pages/Itinerary/LoadingPage";
 
 import { Provider } from "react-redux";
 import store from "./Redux/itinerary_reducer";
