+++ src/pages/Destination/ResultPage.js
@@ -1,5 +1,5 @@
-import Header from "../components/Result/Header";
-import Content from "../components/Result/Content";
+import Header from "../../components/Result/Header";
+import Content from "../../components/Result/Content";
 
 const ResultPage = () => {
   return (
