+++ src/pages/Itinerary/DestinationPage.js
@@ -1,12 +1,12 @@
-import ShortHeader from "../components/Destination/ShortHeader";
-import Title from "../components/Destination/Title";
-import Card from "../components/Destination/Card";
-import Button from "../components/Destination/Button";
-import scrollUp from "../assets/destination/scrollU.png";
-import scrollDown from "../assets/destination/scrollD.png";
+import ShortHeader from "../../components/Destination/ShortHeader";
+import Title from "../../components/Destination/Title";
+import Card from "../../components/Destination/Card";
+import Button from "../../components/Destination/Button";
+import scrollUp from "../../assets/destination/scrollU.png";
+import scrollDown from "../../assets/destination/scrollD.png";
 import { styled } from "styled-components";
 import { useState, useEffect } from "react";
-import { getRegions } from "../api/regions";
+import { getRegions } from "../../api/regions";
 
 const DestinationPage = () => {
   const [destination, setDestination] = useState([]);
@@ -15,41 +15,40 @@ const DestinationPage = () => {
   const [showScrollDownButton, setShowScrollDownButton] = useState(false);
   const [showScrollUpButton, setShowScrollUpButton] = useState(false);
 
-   // 스크롤 다운
+  // 스크롤 다운
   const handleScrollDown = () => {
     setScrollAmount(scrollAmount + 844);
     window.scrollBy(0, 844);
   };
 
-    // 스크롤 업
-    const handleScrollUp = () => {
-      setScrollAmount(scrollAmount - 844);
-      window.scrollBy(0, -844);
+  // 스크롤 업
+  const handleScrollUp = () => {
+    setScrollAmount(scrollAmount - 844);
+    window.scrollBy(0, -844);
+  };
+
+  useEffect(() => {
+    const handleScroll = () => {
+      const windowHeight = window.innerHeight;
+      const documentHeight = document.body.scrollHeight;
+
+      if (window.scrollY === 0) {
+        setShowScrollDownButton(true);
+        setShowScrollUpButton(false);
+      } else if (window.scrollY + windowHeight >= documentHeight) {
+        setShowScrollDownButton(false);
+        setShowScrollUpButton(true);
+      } else {
+        setShowScrollDownButton(true);
+        setShowScrollUpButton(true);
+      }
     };
 
-    useEffect(() => {
-      const handleScroll = () => {
-        const windowHeight = window.innerHeight;
-        const documentHeight = document.body.scrollHeight;
-  
-
-    if (window.scrollY === 0) { 
-      setShowScrollDownButton(true);
-      setShowScrollUpButton(false);
-    }else if (window.scrollY + windowHeight >= documentHeight) {
-          setShowScrollDownButton(false);
-          setShowScrollUpButton(true);
-        } else {
-          setShowScrollDownButton(true);
-          setShowScrollUpButton(true);
-        }
-      };
-  
-      window.addEventListener("scroll", handleScroll);
-      return () => {
-        window.removeEventListener("scroll", handleScroll);
-      };
-    }, []);
+    window.addEventListener("scroll", handleScroll);
+    return () => {
+      window.removeEventListener("scroll", handleScroll);
+    };
+  }, []);
 
   // 지역 데이터 얻기
   const getDestinationData = async () => {
@@ -86,8 +85,12 @@ const DestinationPage = () => {
       </CardContainer>
       <Fixed>
         <ScrollContainer>
-          {showScrollUpButton &&    <ScrollUp onClick={handleScrollUp} src={scrollUp} />}
-        {(scrollAmount === 0 || showScrollDownButton) &&  <ScrollDown  onClick={handleScrollDown} src={scrollDown} />}
+          {showScrollUpButton && (
+            <ScrollUp onClick={handleScrollUp} src={scrollUp} />
+          )}
+          {(scrollAmount === 0 || showScrollDownButton) && (
+            <ScrollDown onClick={handleScrollDown} src={scrollDown} />
+          )}
         </ScrollContainer>
       </Fixed>
       <Button regionId={regionId} />
